% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{cdw}
\alias{cdw}
\title{Shortcuts for interactive functions.}
\usage{
cdw(...)
}
\arguments{
\item{...}{further argument passed to \code{\link[countdown:countdown]{countdown::countdown()}}.}
}
\description{
Shortcuts for interactive functions.
}
