% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{kig}
\alias{kig}
\title{Shortcuts for knitr function}
\usage{
kig(...)
}
\arguments{
\item{...}{further argument passed to \code{\link[knitr:include_graphics]{knitr::include_graphics()}}.}
}
\description{
Shortcuts for knitr function
}
