% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pres.R
\name{new_pres}
\alias{new_pres}
\title{Shortcuts for interactive functions.}
\usage{
new_pres(
  title = "Cool title",
  date = Sys.Date(),
  author = "inSileco Tea",
  pres_name = "newpres",
  overwrite = FALSE
)
}
\arguments{
\item{title}{presentation title.}

\item{date}{presentation date.}

\item{author}{author name.}

\item{pres_name}{name of the folder that includes the presentation.}

\item{overwrite}{a logical. Overwrite files if they exist
(see \code{\link[fs:copy]{fs::dir_copy()}}).}
}
\description{
Shortcuts for interactive functions.
}
