% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{rfa}
\alias{rfa}
\alias{gh}
\alias{gl}
\alias{gt}
\alias{py}
\alias{rp}
\alias{co}
\alias{br}
\alias{db}
\alias{lc}
\alias{ar}
\alias{al}
\alias{bk}
\alias{ch}
\alias{ck}
\alias{nf}
\alias{lk}
\alias{fl}
\alias{fd}
\alias{pd}
\alias{cg}
\alias{cb}
\alias{ey}
\alias{tr}
\alias{rl}
\alias{so}
\alias{wk}
\title{Shortcuts for fontawesome icons.}
\usage{
rfa(..., fill = "currentColor", help = FALSE)

gh()

gl()

gt()

py()

rp()

co()

br()

db()

lc()

ar()

al()

bk()

ch()

ck()

nf()

lk()

fl()

fd()

pd()

cg()

cb()

ey()

tr()

rl()

so()

wk()
}
\arguments{
\item{...}{further argument passed to \code{\link[fontawesome:fa]{fontawesome::fa()}}.}

\item{fill}{color to fill the icon with see \code{\link[fontawesome:fa]{fontawesome::fa()}}.}

\item{help}{if \code{TRUE} then a table with shortcut names (\code{fun}) and
corresponding icons is displayed.}
}
\description{
Shortcuts for fontawesome icons.
}
\references{
See https://fontawesome.com/search
}
