% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wksh_dl.R
\name{wksh_dl}
\alias{wksh_dl}
\title{Download workshops}
\usage{
wksh_dl(id, path = ".", verbose = TRUE)
}
\arguments{
\item{id}{a vector of identifiers.}

\item{path}{path where the workshops files are or a path to a folder where the files will be extracted if download is set to \code{TRUE}.}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Download a selection of workshops.
}
\examples{
\dontrun{
 wksh_dl(1)
 wksh_dl(c(1, 2))
 wksh_dl("a")
}
}
