% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{rfa}
\alias{rfa}
\alias{gh}
\alias{gl}
\alias{gt}
\alias{rp}
\alias{ch}
\alias{ck}
\alias{db}
\alias{nf}
\alias{lk}
\alias{bk}
\alias{fl}
\alias{fd}
\alias{pd}
\alias{ar}
\alias{cg}
\alias{tr}
\alias{cb}
\alias{ey}
\alias{lc}
\alias{so}
\alias{wk}
\title{Shortcuts for fontawesome icons.}
\usage{
rfa(..., fill = "currentColor", help = FALSE)

gh()

gl()

gt()

rp()

ch()

ck()

db()

nf()

lk()

bk()

fl()

fd()

pd()

ar()

cg()

tr()

cb()

ey()

lc()

so()

wk()
}
\arguments{
\item{...}{further argument passed to \code{\link[icons:fontawesome]{icons::fontawesome()}}.}

\item{fill}{color to fill the icon with see \code{\link[icons:icon_style]{icons::icon_style()}}.}

\item{help}{if \code{TRUE} then a table with shortcut names (\code{fun}) and corresponding icons is displayed.}
}
\description{
Shortcuts for fontawesome icons.
}
